+++ Modules/38-StorageManager.ps1
@@ -1755,6 +1755,7 @@ function Show-StorageManagerMenu {
 # Function to run Storage Manager
 function Start-StorageManager {
     while ($true) {
+        if ($global:ReturnToMainMenu) { return }
         $choice = Show-StorageManagerMenu
 
         $navResult = Test-NavigationCommand -UserInput $choice